+++ packages/fluentui/react-northstar/src/themes/teams/components/Toolbar/toolbarMenuItemStyles.ts
@@ -3,7 +3,6 @@ import { getColorScheme } from '../../colors';
 import { getBorderFocusStyles } from '../../getBorderFocusStyles';
 import { ToolbarMenuItemStylesProps } from '../../../../components/Toolbar/ToolbarMenuItem';
 import { ToolbarVariables } from './toolbarVariables';
-import { submenuIndicatorUrl } from './submenuIndicatorUrl';
 import { activeIndicatorUrl } from './activeIndicatorUrl';
 import { toolbarMenuItemSubmenuIndicatorClassName } from '../../../../components/Toolbar/ToolbarMenuItemSubmenuIndicator';
 import { toolbarMenuItemActiveIndicatorClassName } from '../../../../components/Toolbar/ToolbarMenuItemActiveIndicator';
@@ -40,7 +39,7 @@ export const toolbarMenuItemStyles: ComponentSlotStylesPrepared<ToolbarMenuItemS
         color: v.menuItemForegroundHover || colors.menuItemForegroundHover,
         backgroundColor: v.menuItemBackgroundHover || colors.menuItemBackgroundHover,
         [`& .${toolbarMenuItemSubmenuIndicatorClassName}`]: {
-          backgroundImage: submenuIndicatorUrl(v.menuItemForegroundHover || colors.menuItemForegroundHover),
+          color: v.menuItemForegroundHover || colors.menuItemForegroundHover,
         },
         [`& .${toolbarMenuItemActiveIndicatorClassName}`]: {
           backgroundImage: activeIndicatorUrl(v.menuItemForegroundHover || colors.menuItemForegroundHover),
@@ -54,7 +53,7 @@ export const toolbarMenuItemStyles: ComponentSlotStylesPrepared<ToolbarMenuItemS
         color: v.menuItemForegroundDisabled || colors.foregroundDisabled1,
         backgroundColor: v.menuItemBackgroundDisabled,
         [`& .${toolbarMenuItemSubmenuIndicatorClassName}`]: {
-          backgroundImage: submenuIndicatorUrl(v.menuItemForegroundDisabled || colors.foregroundDisabled1),
+          color: v.menuItemForegroundDisabled || colors.foregroundDisabled1,
         },
         [`& .${toolbarMenuItemActiveIndicatorClassName}`]: {
           backgroundImage: activeIndicatorUrl(v.menuItemForegroundDisabled || colors.foregroundDisabled1),